+++ src/Sample.Presentation/Entities/GeoViewPort.cs
@@ -1,0 +1,15 @@
+﻿using Cartography.DynamicMap;
+
+namespace Sample.Entities;
+
+public class GeoViewPort
+{
+	public Geopoint Geopoint;
+	public ZoomLevel ZoomLevel;
+
+	public GeoViewPort(Geopoint geopoint, ZoomLevel zoomLevel)
+	{
+		Geopoint = geopoint;
+		ZoomLevel = zoomLevel;
+	}
+}