+++ src/Sample.Shared.Views/Content/Cartography/StaticMapPage.xaml.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.UI.Xaml.Controls;
+
+namespace Sample;
+
+/// <summary>
+/// An empty page that can be used on its own or navigated to within a Frame.
+/// </summary>
+public sealed partial class StaticMapPage : Page
+{
+    public StaticMapPage()
+    {      
+        this.InitializeComponent();
+    }
+}