+++ src/Sample.Shared.Views/Framework/Version/VersionProvider.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace Sample;
+
+public sealed partial class VersionProvider : IVersionProvider
+{
+	public string VersionString => Version.ToString();
+}