+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Updated documentation.
 - New folder structure.
+- Updated the sample application.
 
 ### Deprecated